core: hook up g_ifacename to L4Configurations_getInterfaceName
This is quite an ugly hack, but I can't figure out a better way, since
the XML file overrides the command line.

That's quite suboptimal, as there is no quick way to override a
default configuration, which is what config.xml should be.

So we hook up into the function and if there is a value from the
command line, override.

Signed-off-by: Inaky Perez-Gonzalez <inaky@linux.intel.com>