trace: log files are not rotated automatically
Current algorithm was logging on 10 different files, closing one and
starting in the next one when it reached a certain threshold. There
were 10 files that were rotated. When the 10th file was done, it
started again in file #1, overwriting it.

This was, obviously, causing data loss of older log entries.

Removed in favour of using logrotate for managing log file size.

Signed-off-by: Inaky Perez-Gonzalez <inaky@linux.intel.com>