WiMAXDB: acknowledge --{enable,disable}-bindb
This adds support for installing a provided binary database as part of
the build configuration/installation phase. The path of the
pre-generated file defaults to a certain value that can be overriden
with --with-wimax-db-bin. The files WiMAX_DB.bin and WiMAX_Def.bin
will be copied from said file.

This commit only updates the configure.ac and corresponding
Makefile.am; a subsequent commit updates the autoconf and automake
templates.

Signed-off-by: Inaky Perez-Gonzalez <inaky.perez-gonzalez@intel.com>