L5Common: Fix SendIndicationToSubscribers for 64-bit systems
For ListItems in the subscribersList, the 'pData' field is used to hold the
value of the subscriber ID, instead of storing a pointer. Iterator_GetNext()
can not read subscriber IDs directly into an L5_TARGET_ID since this may not
be as large as the size of 'pData'. With this change, the subscriber ID is
first read into a pointer, whose value is later copied into a L5_TARGET_ID.

Signed-off-by: David Ward <david.ward@ll.mit.edu>